move avail_easyblocks to easybuild.framework.easyconfig.tools module to make it available in more contexts
